[orc8r][NetworkProbe] move swagger definitions to a separate service
