+++ lte/cloud/configs/service_registry.yml
@@ -83,3 +83,16 @@ services:
     annotations:
       orc8r.io/obsidian_handlers_path_prefixes: >
         /magma/v1/lte/:network_id/sms,
+
+  nprobe:
+    host: "localhost"
+    port: 9666
+    echo_port: 10088
+    proxy_type: "clientcert"
+    labels:
+      orc8r.io/obsidian_handlers: "true"
+      orc8r.io/swagger_spec: "true"
+    annotations:
+      orc8r.io/obsidian_handlers_path_prefixes: >
+        /magma/v1/lte/:network_id/network_probe_tasks,
+        /magma/v1/lte/:network_id/network_probe_destinations,
