+++ lte/cloud/go/services/lte/obsidian/models/validate.go
@@ -436,11 +436,3 @@ func (m *CellularGatewayPoolRecords) ValidateModel() error {
 	}
 	return nil
 }
-
-func (m *NetworkProbeTask) ValidateModel() error {
-	return m.Validate(strfmt.Default)
-}
-
-func (m *NetworkProbeDestination) ValidateModel() error {
-	return m.Validate(strfmt.Default)
-}
