+++ lte/cloud/go/services/nprobe/obsidian/models/network_probe_task_swaggergen.go
