+++ orc8r/cloud/docker/controller/supervisord.conf
@@ -261,3 +261,11 @@ stdout_logfile=NONE
 stderr_logfile=NONE
 stdout_events_enabled=true
 stderr_events_enabled=true
+
+[program:nprobe]
+command=/usr/bin/envdir /var/opt/magma/envdir /var/opt/magma/bin/nprobe -run_echo_server=true -logtostderr=true -v=0
+autorestart=true
+stdout_logfile=NONE
+stderr_logfile=NONE
+stdout_events_enabled=true
+stderr_events_enabled=true
